修正错误
编译问题：amd_make.m,amd.m
1.amd_make 中 mex编译语句去掉-inline参数（已过时）
2.将 mex 编译的命令名字改为amd_test，避免和 matlab built-in  amd 函数混淆。此处修改了 amd_make.m，amd.m->amd_test.m，amd_demo.m（调用函数由amd()->amd_test()）
3.执行语句不再catch，可以输出多一点的错误信息

数据类型对接错误问题
amd_mex.c,amd.h,amd_internal.h,amd_2.c,
matlab 中mwIndex 和 c 中的 int 类型不一致。改 Int为 size_t

为了方便查错增加的printf内容:
amd_mex.c, amd_order.c, amd_valid.c

+++ Include/amd.h
@@ -36,11 +36,14 @@
 #ifndef AMD_H
 #define AMD_H
 
-int amd_order (		    /* returns 0 if OK, negative value if error */
-    int n,		    /* A is n-by-n.  n must be >= 0. */
-    const int Ap [ ],	    /* column pointers for A, of size n+1 */
-    const int Ai [ ],	    /* row indices of A, of size nz = Ap [n] */
-    int P [ ],		    /* output permutation, of size n */
+/* size_t */
+#include "stddef.h"
+
+size_t amd_order (		    /* returns 0 if OK, negative value if error */
+    size_t n,		    /* A is n-by-n.  n must be >= 0. */
+    const size_t Ap [ ],	    /* column pointers for A, of size n+1 */
+    const size_t Ai [ ],	    /* row indices of A, of size nz = Ap [n] */
+    size_t P [ ],		    /* output permutation, of size n */
     double Control [ ],	    /* input Control settings, of size AMD_CONTROL */
     double Info [ ]	    /* output Info statistics, of size AMD_INFO */
 ) ;
@@ -255,13 +258,13 @@ long amd_l_order (	    /* see above for description of arguments */
  * flagged with the AMD_INVALID error code.
  */ 
 
-int amd_preprocess
+size_t amd_preprocess
 (
-    int n,
-    const int Ap [ ],
-    const int Ai [ ],
-    int Rp [ ],
-    int Ri [ ]
+    size_t n,
+    const size_t Ap [ ],
+    const size_t Ai [ ],
+    size_t Rp [ ],
+    size_t Ri [ ]
 ) ;
 
 long amd_l_preprocess
